requirements.txt
python==3.11
pytorch
pytorch-cuda==12.1
faiss-gpu
tokenizers>=0.19.0,<0.20.0
transformers>=4.36.0
sentence-transformers>=2.3.0
huggingface_hub==0.21.0
accelerate
scikit-learn
pandas
tqdm
tensorboard
numpy
dexter-cqa
python-dotenv
matplotlib
nltk
